Refactor story file imports to use components barrel file so design tokens CSS is included
